PasswordManager/requirements.txt
flask 
python-dotenv 
flask-login 
flask_sqlalchemy
cryptography